+++ test/ustore/test_head_version.cc
@@ -33,20 +33,20 @@ constexpr char test_head_version_log[] = "test_head_version.log";
 
 TEST(Worker, HeadVersion_PutBranch) {
   EXPECT_FALSE(head_ver.Exists(key[0], branch[0]));
-  EXPECT_EQ(0, head_ver.ListBranch(key[0]).size());
+  EXPECT_EQ(size_t(0), head_ver.ListBranch(key[0]).size());
 
   head_ver.PutBranch(key[0], branch[0], ver[0]);
   EXPECT_TRUE(head_ver.Exists(key[0], branch[0]));
-  EXPECT_EQ(1, head_ver.ListBranch(key[0]).size());
+  EXPECT_EQ(size_t(1), head_ver.ListBranch(key[0]).size());
   EXPECT_TRUE(head_ver.IsBranchHead(key[0], branch[0], ver[0]));
 
   head_ver.PutBranch(key[0], branch[0], ver[1]);
   EXPECT_FALSE(head_ver.IsBranchHead(key[0], branch[0], ver[0]));
   EXPECT_TRUE(head_ver.IsBranchHead(key[0], branch[0], ver[1]));
-  EXPECT_EQ(1, head_ver.ListBranch(key[0]).size());
+  EXPECT_EQ(size_t(1), head_ver.ListBranch(key[0]).size());
 
   head_ver.PutBranch(key[0], branch[1], ver[2]);
-  EXPECT_EQ(2, head_ver.ListBranch(key[0]).size());
+  EXPECT_EQ(size_t(2), head_ver.ListBranch(key[0]).size());
   EXPECT_TRUE(head_ver.Exists(key[0], branch[1]));
 
   head_ver.PutBranch(key[0], branch[2], ver[3]);
@@ -58,7 +58,7 @@ TEST(Worker, HeadVersion_PutBranch) {
 TEST(Worker, HeadVersion_Load) {
   HeadVersion new_head_ver;
   new_head_ver.LoadBranchVersion(test_head_version_log);
-  EXPECT_EQ(3, new_head_ver.ListBranch(key[0]).size());
+  EXPECT_EQ(size_t(3), new_head_ver.ListBranch(key[0]).size());
   EXPECT_EQ(ver[1], new_head_ver.GetBranch(key[0], branch[0]));
   EXPECT_EQ(ver[2], new_head_ver.GetBranch(key[0], branch[1]));
   EXPECT_EQ(ver[3], new_head_ver.GetBranch(key[0], branch[2]));
@@ -66,7 +66,7 @@ TEST(Worker, HeadVersion_Load) {
 }
 
 TEST(Worker, HeadVersion_GetBranch) {
-  EXPECT_EQ(3, head_ver.ListBranch(key[0]).size());
+  EXPECT_EQ(size_t(3), head_ver.ListBranch(key[0]).size());
   EXPECT_EQ(ver[1], head_ver.GetBranch(key[0], branch[0]));
   EXPECT_EQ(ver[2], head_ver.GetBranch(key[0], branch[1]));
   EXPECT_EQ(ver[3], head_ver.GetBranch(key[0], branch[2]));
@@ -88,8 +88,8 @@ TEST(Worker, HeadVersion_PutLatest) {
 }
 
 TEST(Worker, HeadVersion_GetLatest) {
-  EXPECT_EQ(0, head_ver.GetLatest(key[0]).size());
-  EXPECT_EQ(1, head_ver.GetLatest(key[1]).size());
+  EXPECT_EQ(size_t(0), head_ver.GetLatest(key[0]).size());
+  EXPECT_EQ(size_t(1), head_ver.GetLatest(key[1]).size());
 }
 
 TEST(Worker, HeadVersion_RenameBranch) {
@@ -100,9 +100,9 @@ TEST(Worker, HeadVersion_RenameBranch) {
 }
 
 TEST(Worker, HeadVersion_RemoveBranch) {
-  EXPECT_EQ(3, head_ver.ListBranch(key[0]).size());
+  EXPECT_EQ(size_t(3), head_ver.ListBranch(key[0]).size());
 
   head_ver.RemoveBranch(key[0], branch[2]);
   EXPECT_FALSE(head_ver.Exists(key[0], branch[2]));
-  EXPECT_EQ(2, head_ver.ListBranch(key[0]).size());
+  EXPECT_EQ(size_t(2), head_ver.ListBranch(key[0]).size());
 }
